backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
xmlschema==3.3.2
lxml==5.3.0
pydantic==2.8.2
python-dotenv==1.0.1